requirements.txt
astropy>=5.1
Pillow>=9.2.0
FreeSimpleGUI>=5.0.0